+++ CHANGES.md
@@ -2,10 +2,13 @@
 
 ### ? - ?
 
+##### Breaking Changes :mega:
+
+- `Cesium3DTilesSelection::ViewState` can be supplied a delegate that computes a general tile error measure. The interface for setting a fixed geometric error in the `ViewState` object has been removed; that functionality can be implemented instead using the delegate.
+
 ##### Additions :tada:
 
 - The `CesiumVectorOverlays::VectorTilesRasterOverlay` now supports styling via specifying a `VectorStylingProvider` through the `pStylingProvider` parameter in `VectorTilesRasterOverlayOptions`.
-- `Cesium3DTilesSelection::ViewState` can be supplied an object that computes tile error measure.
 
 ### v0.64.0 - 2026-09-01